+++ apps/dv-admin/src/views/dashboard/components/deposits/Deposits.vue
@@ -5,14 +5,11 @@
 	import { formatDollars } from "@shared/utils/helpers/general";
 	import { useI18n } from "vue-i18n";
 	import { computed } from "vue";
-	import { UiButton, UiTable } from "@dv.net/ui-kit";
+	import { UiIcon, UiTable } from "@dv.net/ui-kit";
 	import { getReplenishmentDeclension, getDate } from "@dv-admin/utils/helpers/dashboard";
-	import { getCurrentCoin } from "@shared/utils/helpers/general";
 	import type { UiTableHeader } from "@dv.net/ui-kit/dist/components/UiTable/types";
-	import { useGeneralStore } from "@dv-admin/stores/general";
 
 	const { depositSummary, isLoadingDeposit } = storeToRefs(useDashboardStore());
-	const { dictionary } = storeToRefs(useGeneralStore())
 	const { t, locale } = useI18n();
 
 	const isVisibleTable = computed<boolean>(() => {
@@ -33,23 +30,19 @@
 			width: "100",
 			columnClass: (row: any) => (row.isMoreDetails ? "column-alignment" : "")
 		},
-		{ name: "transactions_count", label: t("Details and number of top ups") },
 		{
-			name: "details",
+			name: "transactions_count",
+			label: t("Count"),
 			width: "180",
+			columnClass: (row: any) => (row.isMoreDetails ? "column-alignment" : "")
+		},
+		{ name: "details", label: t("Top-Ups") },
+		{
+			name: "arrow",
+			width: "60",
 			columnClass: (row: any) => (row.isMoreDetails ? "column-alignment" : "")
 		}
 	]);
-
-	const isShowOnlyCoin = (currencyId: string): boolean => {
-		const currencies = dictionary.value?.available_currencies;
-		if (!currencies?.length) return false;
-		const targetCoin = getCurrentCoin(currencyId);
-		const matchingCoins = currencies
-			.map(({ id }) => getCurrentCoin(id))
-			.filter((coin) => coin === targetCoin);
-		return matchingCoins.length === 1;
-	};
 </script>
 
 <template>
@@ -61,7 +54,7 @@
 			table-layout="fixed"
 			class="deposit__table"
 			@rowClick="(row) => (row.isMoreDetails = !row.isMoreDetails)"
-			:row-class="(row) => (row.details_by_currency.length > 3 ? '' : 'no-pointer')"
+			:row-class="(row) => (row.details_by_currency.length > 4 ? '' : 'no-pointer')"
 		>
 			<template #body-cell-date="{ row }">
 				{{
@@ -70,50 +63,31 @@
 						: getDate(row.date, row.type, locale).text
 				}}
 			</template>
-
 			<template #body-cell-sum_usd="{ row }"> {{ formatDollars(row.sum_usd) }}</template>
-
 			<template #body-cell-transactions_count="{ row }">
-				<div class="detail">
-					<span class="detail__replenishment">
-						{{ row.transactions_count }} {{ $t(getReplenishmentDeclension(row.transactions_count)) }}
-					</span>
-					<div class="detail__coins" :class="{ hidden: !row.isMoreDetails }">
-						<div
-							class="detail__coin"
-							v-for="item in row.details_by_currency"
-							:key="item.currency"
-						>
-							<span class="detail__coin-blue" :class="{ 'detail__coin-green': item.percentage > 50 }">
-								{{ item.percentage }}%
-							</span>
-							<span>
-								{{ isShowOnlyCoin(item.currency) ? getCurrentCoin(item.currency) : item.currency.replace(".", " ") }}
-							</span>
-						</div>
+				<span class="replenishment">
+					{{ row.transactions_count }} {{ $t(getReplenishmentDeclension(row.transactions_count)) }}
+				</span>
+			</template>
+			<template #body-cell-details="{ row }">
+				<div class="coins" :class="{ hidden: !row.isMoreDetails }">
+					<div
+						class="coins__item"
+						v-for="item in row.details_by_currency"
+						:key="item.currency"
+						:class="{ 'coins__item--green': item.percentage > 50 }"
+					>
+						<span>{{ item.percentage }}%</span>
+						<span>{{ item.currency }}</span>
 					</div>
 				</div>
 			</template>
-
-			<template #body-cell-details="{ row }">
-				<Transition
-					v-if="row.details_by_currency.length > 3"
-					name="fade"
-					mode="out-in"
-					:duration="{ enter: 10, leave: 10 }"
-					style="margin-left: auto"
-				>
-					<ui-button
-						:key="String(row.isMoreDetails)"
-						:type="row.isMoreDetails ? 'outline-light' : 'outline'"
-						:mode="row.isMoreDetails ? 'neutral' : 'accent'"
-						size="sm"
-						:left-icon-name="row.isMoreDetails ? 'keyboard-arrow-up' : 'keyboard-arrow-down'"
-						left-icon-size="sm"
-					>
-						{{ row.isMoreDetails ? $t("Hide") : $t("Show all") }}
-					</ui-button>
-				</Transition>
+			<template #body-cell-arrow="{ row }">
+				<div v-if="row.details_by_currency.length > 4" class="arrow">
+					<span class="arrow__icon" :class="{ 'arrow__icon--rotated': row.isMoreDetails }">
+						<ui-icon type="400" name="keyboard-arrow-down" color="#A4A5B1" />
+					</span>
+				</div>
 			</template>
 		</ui-table>
 	</block-section>
@@ -125,87 +99,83 @@
 		display: flex;
 		flex-direction: column;
 		gap: 20px;
-
 		&__table {
 			&:deep(.ui-table) {
 				.ui-table__body-row {
+					@media (hover: hover) {
+						&:hover {
+							background-color: unset;
+						}
+					}
 					&:not(.no-pointer) {
 						@extend .pointer;
 					}
-
 					.column-alignment {
 						vertical-align: top;
-
 						.ui-table__body-cell-inner {
 							margin-top: 5px;
 						}
-					}
-
-					&:first-child .ui-table__body-cell {
-						&:first-child {
+						&:nth-child(3) {
 							.ui-table__body-cell-inner {
-								font-weight: 500;
-							}
-						}
-
-						&:nth-child(2) {
-							.ui-table__body-cell-inner {
-								font-weight: 500;
+								margin-top: 0;
 							}
 						}
 					}
 				}
 			}
-
-			.detail {
-				display: flex;
+			.replenishment {
+				flex-shrink: 0;
+				min-width: 100px;
+				max-width: 140px;
+				color: $blue;
+				font-size: 12px;
+				font-weight: 500;
+				padding: 4px 8px;
+				border-radius: 4px;
+				border: 1px solid $grey;
+				background-color: $blue-opacity;
+				word-break: break-word;
+			}
+			.coins {
+				display: grid;
+				grid-template-columns: repeat(4, 1fr);
 				gap: 4px;
-				align-items: center;
-				flex-grow: 1;
-
-				&__replenishment {
+				&.hidden {
+					overflow: hidden;
+					height: 30px;
+				}
+				&__item {
 					flex-shrink: 0;
-					align-self: baseline;
-					min-width: 100px;
+					width: 104px;
+					display: flex;
+					align-items: center;
+					justify-content: center;
+					gap: 4px;
+					border-radius: 6px;
+					padding: 4px 8px;
 					color: $blue;
 					font-size: 12px;
 					font-weight: 500;
-					padding: 4px 8px;
-					border-radius: 4px;
-					border: 1px solid $grey;
-					background-color: $blue-opacity;
-				}
-
-				&__coins {
-					display: flex;
-					flex-wrap: wrap;
-					gap: 4px;
-					flex-grow: 1;
-
-					&.hidden {
-						overflow: hidden;
-						height: 30px;
+					background: #F7F9FB;
+					word-break: break-word;
+					&--green {
+						color: #1F9649;
+						background: rgba(31, 150, 73, 0.08);
 					}
 				}
-
-				&__coin {
-					display: flex;
-					align-items: center;
-					gap: 4px;
-					border-radius: 4px;
-					border: 1px solid $grey;
-					padding: 4px 8px;
-					color: $secondary;
-					font-size: 12px;
-					font-weight: 500;
-					width: max-content;
-
-					&-blue {
-						color: $blue;
-					}
-
-					&-green {
-						color: #5dd565;
+			}
+			.arrow {
+				width: 100%;
+				display: flex;
+				justify-content: flex-end;
+				&__icon {
+					flex-shrink: 0;
+					@extend .center;
+					width: 25px;
+					height: 25px;
+					transition: transform 0.3s ease;
+					&--rotated {
+						transform: rotate(180deg);
 					}
 				}
 			}